Removed not NULL requirement on tlmc_path
Any '' values in tlmc_path were updated to NULL
As well removed any '' values in url_links with NULL

The original tlmc_path not NULL requirement was due to prior different use for that row under another use case. As that has been abandoned in favour for the new direction this has been now patched.