Notice regarding content
Just a minor update affirming that this is not a spot where illegal content should be attempted to be merged and posted. The only links are intended to be to the official pages of the artist.

also minor site template push, no functional change just clarifying the Links and removing some old comment.